+++ lib/services/di/services_dependency_container.dart
@@ -9,6 +9,7 @@ import 'package:zachranobed/services/delivery_service.dart';
 import 'package:zachranobed/services/entity_pairs_service.dart';
 import 'package:zachranobed/services/entity_service.dart';
 import 'package:zachranobed/services/food_box_service.dart';
+import 'package:zachranobed/services/meal_service.dart';
 import 'package:zachranobed/services/offered_food_service.dart';
 import 'package:zachranobed/services/shipping_of_boxes_service.dart';
 
@@ -19,6 +20,7 @@ class ServicesDependencyContainer {
     GetIt.I.registerSingleton(FoodBoxService());
     GetIt.I.registerSingleton(EntityService());
     GetIt.I.registerSingleton(EntityPairService());
+    GetIt.I.registerSingleton(MealService());
     GetIt.I.registerSingleton(
       AuthService(
         GetIt.I<EntityService>(),